fix: Long expected, not Int

+++ app/src/main/java/dev/trindadedev/anims/ui/components/drag/DragState.kt
@@ -3,5 +3,5 @@ package dev.trindadedev.anims.ui.components.drag
 data class DragState(
   var dY: Float = 0f,
   var dX: Float = 0f,
-  var duration: Int = 0
+  var duration: Long = 0
 )
